MPI_Reduce_scatter[_block] fix the receive buffer size.
In the case of MPI_Reduce_scatter(), the sum of all the recvcounts
should be equal to the number of elements in the send buffer.

In the case of MPI_Reduce_scatter_block(), recvcount * comm_size
should be equal to the number of elements in the send buffer.

This issue was initially reported by Simon David Hammond at
https://www.mail-archive.com/users@lists.open-mpi.org/msg32889.html